// src/app/model/loan.ts
//This will be modeling my data from my backend api.
export interface Loan {
  id: number | undefined;
  applicantName: string | undefined;
  phoneNo: string | undefined;
  address: string | undefined;
  email: string | undefined;
  creditScore: number | undefined;
  loanNumber: number | undefined;
  interestRate: number | undefined;
}